backend/requirements.txt
fastapi
uvicorn
google-generativeai
scapy
pandas
supabase
python-multipart
python-dotenv
psutil
virustotal-python
